+++ aqua/core/intake_drivers/xarray/__init__.py
@@ -1,0 +1,5 @@
+from .compat import install_intake_xarray_stub
+from .netcdf import IntakeNetCDFSource
+from .xzarr import IntakeZarrSource
+
+__all__ = ["IntakeNetCDFSource", "IntakeZarrSource", "install_intake_xarray_stub"]